+++ src/main/java/web/service/UserService.java
@@ -10,10 +10,10 @@ public interface UserService {
 
     void updateUser(User updateUser);
 
-    void removeUserById(int id);
+    void removeUserById(Long id);
 
     List<User> getAllUsers();
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
 }
